requirements.txt
acres==0.2.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
comm==0.2.2
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.1
debugpy==1.8.12
decorator==5.1.1
deepsearch-glm==1.0.0
deprecation==2.1.0
dill==0.3.9
distro==1.9.0
docling==2.22.0
docling-core==2.18.1
docling-ibm-models==3.3.2
docling-parse==3.3.1
easyocr==1.7.2
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.2.0
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
filetype==1.2.0
fitz==0.0.1.dev2
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-genai==1.2.0
gradio==5.16.0
gradio_client==1.7.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.32.0
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lancedb==0.19.0
latex2mathml==3.77.0
lazy_loader==0.4
looseversion==1.3.0
lxml==5.3.1
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.17
narwhals==1.26.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.3
nipype==1.9.2
numpy==2.2.3
openai==1.63.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pdfid==1.1.3
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==5.29.3
prov==2.0.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
pydub==0.25.1
Pygments==2.19.1
pylance==0.23.0
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.2.1
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
ruff==0.9.6
safehttpx==0.1.6
safetensors==0.5.2
scikit-image==0.25.1
scipy==1.15.1
semantic-version==2.10.0
semchunk==2.2.2
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.45.3
streamlit==1.42.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tifffile==2025.1.10
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.48.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==14.2
XlsxWriter==3.2.2
